批量删除时，没有选择删除项的BUG

+++ src/main/webapp/WEB-INF/views/index.jsp
@@ -1,5 +1,5 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8" %>
-<%@taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
 <html>
 <head>
@@ -628,17 +628,22 @@
         empNames = empNames.substring(0,empNames.length-1);
         //去除del_idstr字符串中多余的“-”
         del_idstr = del_idstr.substring(0,del_idstr.length-1);
-        if(confirm("确认删除【"+empNames+"】吗？")){
-            //发送ajax请求进行删除
-            $.ajax({
-                url:"${APP_PATH}/emp/"+del_idstr,
-                type:"DELETE",
-                success:function (result) {
-                    alert(result.msg);
-                    //回到当前页面
-                    to_page(currentPage);
-                }
-            })
+        // if(typeof empNames !== "undefined" || empNames !== null || empNames !== ""){
+        if(empNames != ""){
+            if(confirm("确认删除【"+empNames+"】吗？")){
+                //发送ajax请求进行删除
+                $.ajax({
+                    url:"${APP_PATH}/emp/"+del_idstr,
+                    type:"DELETE",
+                    success:function (result) {
+                        alert(result.msg);
+                        //回到当前页面
+                        to_page(currentPage);
+                    }
+                })
+            }
+        }else{
+           alert("请至少选择一个员工进行删除！")
         }
     })
 </script>
